+++ packages/amplify-graphql-transformer-test-utils/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.6](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.4.5...@aws-amplify/graphql-transformer-test-utils@0.4.6) (2024-02-28)
+
+**Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils
+
 ## [0.4.5](https://github.com/aws-amplify/amplify-category-api/compare/@aws-amplify/graphql-transformer-test-utils@0.4.4...@aws-amplify/graphql-transformer-test-utils@0.4.5) (2024-02-05)
 
 **Note:** Version bump only for package @aws-amplify/graphql-transformer-test-utils